requirements.txt
cv2
pillow==6.2.1

torch>=1.1
torchvision
numpy

pandas
scikit-learn 

tqdm
tensorboard>=1.14

efficientnet

easydict